+++ buckchat/app/index.js
@@ -2,10 +2,7 @@
 
 var winston       = require('winston')
   , blueprint     = require('@onehilltech/blueprint')
-  , express       = require('express')
   ;
-
-var expressApp = express();
 
 blueprint.Application (__dirname, function (err, app) {
   if (err) throw err;
