+++ buckchat/app/listeners/app.init/initPassport.js
@@ -1,3 +1,6 @@
+// This listener is copied from Blueprint.js passport authentication tutorial 
+// at https://github.com/onehilltech/blueprint/wiki/Tutorials%3AAuthentication%3ASession
+
 var passport      = require ('passport')
   , LocalStrategy = require ('passport-local').Strategy
   ;
@@ -11,11 +14,11 @@ function initPassport (app) {
   passport.use (new LocalStrategy (opts, authorize));
 
   function authorize (username, password, done) {
-    console.log("Authorizing user!!!!!");
+    console.log("Authenticating user....");
     User.findOne ({ username: username }, function (err, user) {
       if (err) { return done (err); }
       if (!user) { return done (null, false); }
-      if (user.verifyPassword (password)) { return done (null, false); }
+      if (user.password != password) { return done (null, false); }
       return done (null, user);
     });
   }
