--- buckchat/app/routers/DashRouter.js
@@ -1,16 +1,0 @@
-function isLoggedIn (req, res, next) {
-  // if user is authenticated in the session, carry on
-  if (req.isAuthenticated())
-    return next ();
-
-  // if they aren't redirect them to the login page
-  res.redirect ('/login');
-}
-
-module.exports = {
-  '/dashboard': {
-    use: isLoggedIn,
-
-    // rest of definition
-  }
-};
