+++ buckchat/app/routers/LoginRouter.js
@@ -3,8 +3,9 @@ var passport = require ('passport')
 
 module.exports = {
   '/login' : {
-    get  : { view   : 'intro.pug' },
+    get  : { view : 'intro.pug' },
     post : {
+      // Use passport to handle authentication before invoking the `login` function in the controller. 
       before : [passport.authenticate('local', {failureRedirect: '/login'})],
       action : 'LoginController@login'
     }
